+++ packages/go/badge_settings/types.go
@@ -2,6 +2,8 @@
 
 package badge_settings
 
+import "encoding/json"
+
 type BadgeSettingsCaptureReward struct{}
 
 type BadgeSettingsCaptureRewardV2 struct{}
@@ -20,7 +22,7 @@ type BadgeSettingsEventBadgeSettings struct {
 
 type BadgeSettings struct {
 	BadgeRank          uint64                           `json:"badgeRank"`
-	BadgeType          string                           `json:"badgeType"`
+	BadgeType          json.RawMessage                  `json:"badgeType"`
 	CaptureReward      []any                            `json:"captureReward,omitempty"`
 	EventBadge         *bool                            `json:"eventBadge,omitempty"`
 	EventBadgeSettings *BadgeSettingsEventBadgeSettings `json:"eventBadgeSettings,omitempty"`